bugfix - some changes after first feedback
