+++ CNF/appCNF.js
@@ -22,7 +22,7 @@ createApp({
                     return;
                 }
                 const parsed = JSON.parse(rawData);
-                console.log("Parsed CNF data:", parsed);
+                //console.log("Parsed CNF data:", parsed);
                 if (parsed.length === 0) {
                     alert(errorMessages.NO_CNF_STEPS_ERROR);
                     window.history.back();
