+++ pages/manual_page.html
@@ -10,8 +10,6 @@
     <link rel="stylesheet" href="https://unpkg.com/mathlive/dist/mathlive.css">
     <script src="https://unpkg.com/mathlive"></script>
 
-    <!-- Vue 3 -->
-    <script src="https://unpkg.com/vue@3/dist/vue.global.prod.js"></script>
 
     <!-- Styles -->
     <link rel="stylesheet" href="../styles/manual_app.css" />
@@ -57,7 +55,7 @@
                 <li><code>/^[a-r][0-9]*$/</code> Constant symbols, to represent specific objects of domain you can use only lowercase letters from "a" to "r" with numbers after them, for example: <math-field class="static-formula" read-only="true">{b, a1}</math-field></li>
                 <li><code>/^[s-z][0-9]*$/</code> Variable symbols, to represent unknown objects of domain you can use only lowercase letters from "s" to "z" with numbers after them, for example: <math-field class="static-formula" read-only="true">{v, z1}</math-field></li>
                 <li>To represent quantified variables you cannot use parentheses before predicates, for example: <math-field class="static-formula" read-only="true">{\left(\forall x P(x)\right)}</math-field></li>
-                <li>Since the project uses MathLive input field to go to a new line you will need to use special construction (if you would like to paste multiple formulas) or just use <code>Ctrl+Enter<\code>, for example: <math-field class="static-formula" read-only="true">\displaylines{P(a,b)\\ O(a)}</math-field></li>
+                <li>Since the project uses MathLive input field to go to a new line you will need to use special construction (if you would like to paste multiple formulas) or just use <code>Ctrl+Enter</code>, for example: <math-field class="static-formula" read-only="true">\displaylines{P(a,b)\\ O(a)}</math-field></li>
             </ul>
 
             <p>Examples of input formulae you can find on the <a href="https://fuveluck.github.io/">main page</a> in navigation bar :)</p>
